+++ dir.c
@@ -248,8 +248,8 @@ struct directory_contents
      * qualified name of the directory. Beware though, this is also
      * unreliable. I'm open to suggestion on a better way to emulate inode.  */
     char *path_key;
-    int   ctime;
-    int   mtime;        /* controls check for stale directory cache */
+    long long int   ctime;
+	long long int   mtime;        /* controls check for stale directory cache */
     int   fs_flags;     /* FS_FAT, FS_NTFS, ... */
 # define FS_FAT      0x1
 # define FS_NTFS     0x2
@@ -405,8 +405,8 @@ static unsigned int open_directories = 0;
 struct dirfile
   {
     const char *name;           /* Name of the file.  */
-    short length;
-    short impossible;           /* This file is impossible.  */
+    int length;
+    int impossible;           /* This file is impossible.  */
   };
 
 static unsigned long
